+++ finagle-core/src/main/scala/com/twitter/finagle/loadbalancer/EndpointFactory.scala
@@ -3,7 +3,8 @@ package com.twitter.finagle.loadbalancer
 import com.twitter.finagle._
 import com.twitter.finagle.addr.WeightedAddress
 import com.twitter.finagle.service.FailingFactory
-import com.twitter.util.{Future, Time}
+import com.twitter.util.Future
+import com.twitter.util.Time
 import java.util.concurrent.atomic.AtomicReference
 import scala.annotation.tailrec
 import scala.util.control.NonFatal
@@ -20,7 +21,7 @@ trait EndpointFactory[Req, Rep] extends ServiceFactory[Req, Rep] {
    */
   def address: Address
 
-  private[loadbalancer] lazy val weight: Double = WeightedAddress.extract(address)._2
+  private[loadbalancer] def weight: Double = WeightedAddress.extract(address)._2
 
   /**
    * Signals to the endpoint that it should close and rebuild