+++ finagle-core/src/main/scala/com/twitter/finagle/loadbalancer/LoadBalancerFactory.scala
@@ -2,13 +2,17 @@ package com.twitter.finagle.loadbalancer
 
 import com.twitter.finagle._
 import com.twitter.finagle.client.Transporter
-import com.twitter.finagle.loadbalancer.aperture.{EagerConnections, WeightedApertureToggle}
-import com.twitter.finagle.loadbalancer.distributor.AddressedFactory
+import com.twitter.finagle.loadbalancer.aperture.EagerConnections
+import com.twitter.finagle.loadbalancer.aperture.WeightedApertureToggle
 import com.twitter.finagle.service.FailFastFactory
 import com.twitter.finagle.stats._
-import com.twitter.finagle.util.{DefaultLogger, DefaultMonitor}
-import com.twitter.util.{Activity, Event, Var}
-import java.util.logging.{Level, Logger}
+import com.twitter.finagle.util.DefaultLogger
+import com.twitter.finagle.util.DefaultMonitor
+import com.twitter.util.Activity
+import com.twitter.util.Event
+import com.twitter.util.Var
+import java.util.logging.Level
+import java.util.logging.Logger
 import com.twitter.finagle.loadbalancer.distributor.AddrLifecycle
 import scala.util.control.NonFatal
 
@@ -83,11 +87,11 @@ object LoadBalancerFactory {
    * If this is configured, the [[Dest]] param will be ignored.
    */
   private[finagle] case class Endpoints(
-    va: Event[Activity.State[Set[AddressedFactory[_, _]]]])
+    va: Event[Activity.State[Set[EndpointFactory[_, _]]]])
 
   private[finagle] object Endpoints {
     implicit val param =
-      Stack.Param(Endpoints(Event[Activity.State[Set[AddressedFactory[_, _]]]]()))
+      Stack.Param(Endpoints(Event[Activity.State[Set[EndpointFactory[_, _]]]]()))
   }
 
   /**
@@ -370,7 +374,7 @@ object LoadBalancerFactory {
       // cluster to another, and crucially, to share data between endpoints
       val endpoints = if (params.contains[LoadBalancerFactory.Endpoints]) {
         params[LoadBalancerFactory.Endpoints].va
-          .asInstanceOf[Event[Activity.State[Set[AddressedFactory[Req, Rep]]]]]
+          .asInstanceOf[Event[Activity.State[Set[EndpointFactory[Req, Rep]]]]]
       } else {
         TrafficDistributor.weightEndpoints(
           AddrLifecycle.varAddrToActivity(dest, label),
@@ -383,18 +387,10 @@ object LoadBalancerFactory {
       // newBalancer in a TrafficDistributor.
       if (loadBalancerFactory.supportsWeighted && WeightedApertureToggle(label)) {
 
-        // Convert endpoints from AddressedFactories to EndpointFactories
-        val formattedEndpoints: Activity[Set[EndpointFactory[Req, Rep]]] = {
-          Activity(endpoints).map { set: Set[AddressedFactory[Req, Rep]] =>
-            set.map { af: AddressedFactory[Req, Rep] =>
-              af.factory
-            }
-          }
-        }
         // Add the newBalancer to the stack
         Stack.leaf(
           role,
-          newBalancer(formattedEndpoints, disableEagerConnections = false, manageWeights = true)
+          newBalancer(Activity(endpoints), disableEagerConnections = false, manageWeights = true)
         )
       } else {
         // Instead of simply creating a newBalancer here, we defer to the